Warnings removed.

+++ Persistent_Data/CRC16.h
@@ -9,22 +9,22 @@ extern ROM uint16_t CRCtable[];
 
 static inline uint16_t CRC16( const uint16_t input, uint16_t crc)
 {
-    uint8_t carry=(crc >> 8) ^ input;
+    uint8_t carry=(uint8_t)((crc >> 8) ^ input);
     crc = (crc << 8) ^ CRCtable[carry];
 	return crc;
 }
 
-static inline uint16_t CRC16_blockcheck_bytes( const uint8_t *input, uint16_t length)
+static inline uint16_t CRC16_blockcheck_bytes( const uint8_t *input, unsigned  length)
 {
   uint16_t crc=0;
   while( length--)
     {
-      crc = CRC16( (const uint16_t)(*input++), crc);
+      crc = CRC16( (uint16_t)(*input++), crc);
     }
   return crc;
 }
 
-static inline uint16_t CRC16_blockcheck( const uint16_t *input, uint16_t length)
+static inline uint16_t CRC16_blockcheck( const uint16_t *input, unsigned length)
 {
   uint16_t crc=0;
   while( length--)
